Fix to incorrect tint of back & close icon after changing tool sets
If the last window drawn before dropping to the command line had the input focus then a *TOOLSPRITES command is used the default translation table would be tinted.
On reentering the desktop the back & close icon of the first window draw (= top of stack) would be the wrong colour.

Tagged as Wimp-5_30-pre3.