Fix for incorrect names on pinned windows
The post increment of R14 when looking for a control/space terminator meant that the search for a leafname started one too far along.
So in the situation where memory contained something like
  <window_title><null><dot><garbage>
the code would pick up the dot and try to use <garbage> as a leafname.
Submission from Fred Graute.
Fix for ticket #244.

Version 5.42. Tagged as 'Wimp-5_42'