Add ptr_drop to the sprite pool
This is a left handed pointer to signify a drag and drop operation in progress.
The sprite makes use of the mask to define the hotspot to be the tip of the arrow, for Wimp/SpriteExtends that understand that.

Propagated the hotspot definition on ptr_write to the other sprite sets, previously it was only in the Iyonix theme.
Missing ptr_menu and friends added to Morris theme.
Made file_ffb more distinct from the other system files in Ursula theme.

Version 5.47. Tagged as 'Wimp-5_47'