Fix for crashes loading sprites on RISC OS 3.10
A register fumble (revision 4.43 of this file) meant the pointer to the error block from OS_ReadSysInfo not existing was being cached as IRQsema, so later any attempts to resize the RMA would be skipped believing it to be unsafe (because the error pointer would point to an error block whose first word is mostly non zero).
Tested on a real A5000 and emulated via ArcEm.