Fix delivery of extended scroll requests
Extended scroll requests and regular scroll requests share the same
reason code, but are different sizes. Update calcmessagesize to
calculate the size of extended scroll messages correctly, so that the
icon number won't get chopped off when the message is being passed to
the target task.

Version 5.73. Tagged as 'Wimp-5_73'