Fix Message_TaskInitialise reporting wrong slot size
Message_TaskInitialise was reporting that MemoryLimit was the slot size
of the application, which was resulting in the task manager showing the
wrong slot size if for some reason MemoryLimit had been artificially
reduced below ApplicationSpaceSize (e.g. due to the task being a child
of a C system() call). And because MemoryLimit is lower than
ApplicationSpaceSize, any attempt the task makes to grow/shrink its wimp
slot is denied, so task manager will be stuck showing the wrong value
for the entire lifetime of the task.

Fix Message_TaskInitialise to use ApplicationSpaceSize instead, so that
the total size of the slot is reported instead of just the usable area.

This fixes task manager incorrectly showing that Fat32Fs has a slot size
of 0 on some of my Raspberry Pis, when really it has a size of 256K.
This discrepancy was also causing the "System memory allocation" numbers
to not total up correctly (the sum was 256K lower than the displayed
total).