[509] Fix abort autoscrolling off a window with no icons
When looking up the user icon button type passed in R3/R4, use the window handle
in R3 rather than 'handle' which is the window in which autoscrolling started.
This caused an abort if that window contained no icons (because w_icons = -1, and
R4 != -1 leading to R5 = 15 and invariably an address in zero page somewhere).