Rewrite rules for resources
BuildSys 7.91 changes how top-level phony rules are overridden.
In this component, the `resources` target is redefined after `AAsmModule`
is included, so it will be affected. Rewrite the rules so that they will
continue to work the same (at least for native builds) both before and
after BuildSys is updated.

Version 5.86. Not tagged