[581] Revise assumption that ROM toolsprites are always unmasked
Wimp-3_99o introduced an assumption that any toolsprites in ROM are to be
plotted unmasked, and that holds for Ursula and later where the ROM set don't
need the underlying colour (gadgets are always grey and rely on the tintfunc to
make the focus titlebar cream).
The toolsprites in the RISC OS 3.10 ROM however are almost entirely transparent
as their 2D look just plots pixels on top of the underlying fill colour.

Change the logic to look at all the toolsprites and only use the solid plot
SpriteOp if they are all truly solid (skipping the titlebar, since those are
already special cased by dofunkytitlebar and the tintfunc).