Make use of button well colour option for Raspberry theme
Wimp 5.81 and later support a set of WimpVisualFlags which allow the button
colours to be customised per theme. While the faces and shadow in the
Raspberry theme have a similar tone to tile_1, the yellow well around a default
action button jars with the focus window title bar yellow.
Use a -ButtonWellColour switch to match them.


Version 5.87. Not tagged